Clean up test setup Add test/mockStore, which builds a mock store with a branching workflow, tasks and a mock client. Use the mock store in tests for task nav, annotated steps and classifications.
