+++ packages/lib-classifier/src/components/Classifier/components/TaskArea/components/Tasks/components/TaskNavButtons/TaskNavButtonsContainer.spec.js
@@ -2,105 +2,19 @@ import React from 'react'
 import { shallow } from 'enzyme'
 import { expect } from 'chai'
 import { Provider } from 'mobx-state-tree'
-import { Factory } from 'rosie'
 import sinon from 'sinon'
 import TaskNavButtonsContainer from './TaskNavButtonsContainer'
 import TaskNavButtons from './TaskNavButtons'
 
-import RootStore from '@store'
 import Step from '@store/WorkflowStepStore/Step'
-import { SubjectFactory, WorkflowFactory, ProjectFactory } from '@test/factories'
-import stubPanoptesJs from '@test/stubPanoptesJs'
+import mockStore from '@test/mockStore'
 
 describe('TaskNavButtonsContainer', function () {
-  function setupMocks () {
-    const singleChoiceTask = {
-      answers: [{ label: 'yes', next: 'T1' }, { label: 'no' }],
-      question: 'Is there a cat?',
-      required: '',
-      taskKey: 'T0',
-      type: 'single'
-    }
-    const multipleChoiceTask = {
-      answers: [{ label: 'napping' }, { label: 'standing' }, { label: 'playing' }],
-      question: 'What is/are the cat(s) doing?',
-      required: '',
-      taskKey: 'T1',
-      type: 'multiple'
-    }
-    const workflowSnapshot = WorkflowFactory.build({
-      id: 'tasksWorkflow',
-      display_name: 'A test workflow',
-      first_task: 'T0',
-      tasks: {
-        T0: singleChoiceTask,
-        T1: multipleChoiceTask
-      },
-      version: '0.0'
-    })
-    const subjectSnapshot = SubjectFactory.build({
-      id: 'subject',
-      metadata: {}
-    })
-    const projectSnapshot = ProjectFactory.build({
-      id: 'project'
-    })
-    const { panoptes } = stubPanoptesJs({
-      field_guides: [],
-      projects: [projectSnapshot],
-      subjects: Factory.buildList('subject', 10),
-      tutorials: [],
-      workflows: [workflowSnapshot]
-    })
-    const client = {
-      caesar: { request: sinon.stub().callsFake(() => Promise.resolve({})) },
-      panoptes,
-      tutorials: {
-        get: sinon.stub().callsFake(() =>
-          Promise.resolve({ body: {
-            tutorials: []
-          }})
-        )
-      }
-    }
-    const rootStore = RootStore.create({
-      projects: {
-        active: projectSnapshot.id,
-        resources: {
-          [projectSnapshot.id]: projectSnapshot
-        }
-      },
-      subjects: {
-        active: subjectSnapshot.id,
-        resources: {
-          [subjectSnapshot.id]: subjectSnapshot
-        }
-      },
-      workflows: {
-        active: workflowSnapshot.id,
-        resources: {
-          [workflowSnapshot.id]: workflowSnapshot
-        }
-      }
-    }, {
-      authClient: {
-        checkBearerToken: sinon.stub().callsFake(() => Promise.resolve(null)),
-        checkCurrent: sinon.stub().callsFake(() => Promise.resolve(null))
-      },
-      client
-    })
-    rootStore.workflows.setResources([workflowSnapshot])
-    rootStore.workflows.setActive(workflowSnapshot.id)
-    rootStore.subjects.setResources([subjectSnapshot])
-    rootStore.subjects.advance()
-    return rootStore
-  }
-
   describe('when it renders', function () {
     let wrapper
 
     before(function () {
-      const classifierStore = setupMocks()
+      const classifierStore = mockStore()
       const store = { classifierStore }
       wrapper = shallow(<TaskNavButtonsContainer store={store} />)
     })
@@ -119,7 +33,7 @@ describe('TaskNavButtonsContainer', function () {
     let classifierStore
 
     beforeEach(function () {
-      classifierStore = setupMocks()
+      classifierStore = mockStore()
       const store = { classifierStore }
       wrapper = shallow(<TaskNavButtonsContainer store={store} />)
     })
@@ -154,7 +68,7 @@ describe('TaskNavButtonsContainer', function () {
     let classifierStore
 
     before(function () {
-      classifierStore = setupMocks()
+      classifierStore = mockStore()
       const store = { classifierStore }
       wrapper = shallow(<TaskNavButtonsContainer store={store} />)
     })
@@ -206,7 +120,7 @@ describe('TaskNavButtonsContainer', function () {
     const preventDefaultSpy = sinon.spy()
 
     before(function () {
-      classifierStore = setupMocks()
+      classifierStore = mockStore()
       const store = { classifierStore }
       sinon.stub(classifierStore.classifications, 'completeClassification')
