+++ packages/lib-classifier/src/store/Classification/Classification.spec.js
@@ -1,103 +1,13 @@
-import { Factory } from 'rosie'
-import sinon from 'sinon'
-
 import taskRegistry from '@plugins/tasks'
 import TranscriptionLine from '@plugins/drawingTools/experimental/models/marks/TranscriptionLine'
 import Point from '@plugins/drawingTools/models/marks/Point'
 import Line from '@plugins/drawingTools/models/marks/Line'
 
-import RootStore from '@store'
 import Classification, { ClassificationMetadata } from './'
-import { SubjectFactory, WorkflowFactory, ProjectFactory } from '@test/factories'
-import stubPanoptesJs from '@test/stubPanoptesJs'
+import mockStore from '@test/mockStore'
 
 describe('Model > Classification', function () {
   let model
-
-  function setupMocks () {
-    const singleChoiceTask = {
-      answers: [{ label: 'yes', next: 'T1' }, { label: 'no' }],
-      question: 'Is there a cat?',
-      required: '',
-      taskKey: 'T0',
-      type: 'single'
-    }
-    const textTask = {
-      instruction: 'type something',
-      taskKey: 'T0',
-      type: 'text'
-    }
-    const workflowSnapshot = WorkflowFactory.build({
-      id: 'tasksWorkflow',
-      display_name: 'A test workflow',
-      first_task: 'T0',
-      tasks: {
-        T0: singleChoiceTask,
-        T1: textTask
-      },
-      version: '0.0'
-    })
-    const subjectSnapshot = SubjectFactory.build({
-      id: 'subject',
-      metadata: {}
-    })
-    const projectSnapshot = ProjectFactory.build({
-      id: 'project'
-    })
-    const { panoptes } = stubPanoptesJs({
-      field_guides: [],
-      projects: [projectSnapshot],
-      subjects: Factory.buildList('subject', 10),
-      tutorials: [],
-      workflows: [workflowSnapshot]
-    })
-    const client = {
-      caesar: { request: sinon.stub().callsFake(() => Promise.resolve({})) },
-      panoptes,
-      tutorials: {
-        get: sinon.stub().callsFake(() =>
-          Promise.resolve({ body: {
-            tutorials: []
-          }})
-        )
-      }
-    }
-    const rootStore = RootStore.create({
-      projects: {
-        active: projectSnapshot.id,
-        resources: {
-          [projectSnapshot.id]: projectSnapshot
-        }
-      },
-      subjects: {
-        active: subjectSnapshot.id,
-        resources: {
-          [subjectSnapshot.id]: subjectSnapshot
-        }
-      },
-      workflows: {
-        active: workflowSnapshot.id,
-        resources: {
-          [workflowSnapshot.id]: workflowSnapshot
-        }
-      }
-    }, {
-      authClient: {
-        checkBearerToken: sinon.stub().callsFake(() => Promise.resolve(null)),
-        checkCurrent: sinon.stub().callsFake(() => Promise.resolve(null))
-      },
-      client
-    })
-    rootStore.workflows.setResources([workflowSnapshot])
-    rootStore.workflows.setActive(workflowSnapshot.id)
-    rootStore.subjects.setResources([subjectSnapshot])
-    rootStore.subjects.advance()
-    // set an answer to the branching task question, so that step.next is set.
-    const classification = rootStore.classifications.active
-    const singleChoiceAnnotation = classification.annotation(singleChoiceTask)
-    singleChoiceAnnotation.update(0)
-    return rootStore
-  }
 
   before(function () {
     model = Classification.create({
@@ -163,16 +73,16 @@ describe('Model > Classification', function () {
     let secondAnnotation
 
     before(function () {
-      const store = setupMocks()
+      const store = mockStore()
       const classification = store.classifications.active
       // lets update a couple of mock annotations
-      let { annotations } = store.annotatedSteps.latest
-      firstAnnotation = annotations[0]
-      firstAnnotation.update(0)
-      store.annotatedSteps.next()
-      annotations = store.annotatedSteps.latest.annotations
-      secondAnnotation = annotations[0]
-      secondAnnotation.update('This is a text task.')
+      let step = store.workflowSteps.active
+      const [ firstTask ] = step.tasks
+      firstAnnotation = classification.addAnnotation(firstTask, 0)
+      store.workflowSteps.selectStep('S1')
+      step = store.workflowSteps.active
+      const [ secondTask ] = step.tasks
+      secondAnnotation = classification.addAnnotation(secondTask, [1])
       snapshot = classification.toSnapshot()
     })
