+++ packages/lib-classifier/test/mockStore/index.js
@@ -1,0 +1,1 @@
+export { default } from './mockStore'
